// components/Stack/stack-types.ts
// types/stack.types.ts

export type AlignType =
  | "stretch"
  | "flex-start"
  | "center"
  | "flex-end"
  | "baseline";

export type JustifyType =
  | "flex-start"
  | "center"
  | "flex-end"
  | "space-between"
  | "space-around"
  | "space-evenly";

export type GapSizeType = "xs" | "sm" | "md" | "lg" | "xl" | number;
// | string;